backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-taggit==5.0.1
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
isort==5.13.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
msgpack==1.0.8
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
setuptools==69.5.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
swagger-spec-validator==3.0.3
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
